Futher abstraction

+++ spaceinv3/src/spaceinv/model/Shooter.java
@@ -12,7 +12,7 @@ public class Shooter {
     // Will create vertical moving projectile starting
     // at the firing objects "top/bottom-center"
     // Handle the projectile over to the "game loop" to move it.
-    public static Projectile fire(Positionable positionable, int dy) {
+    public static Projectile fire(Positionable positionable, double dy) {
         Projectile p = new Projectile(dy);
         p.setX(positionable.getX() + positionable.getWidth() / 2 - p.getWidth() / 2);
         p.setY(positionable.getY() - p.getHeight());
